Convert vp56_mv to 16-bit. Saves nothing except a bit of memory/cache now, but will allow future optimizations.
Originally committed as revision 24411 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavcodec/vp56.h
@@ -62,8 +62,8 @@ typedef struct {
 } VP56RefDc;
 
 struct vp56_mv {
-    int x;
-    int y;
+    int16_t x;
+    int16_t y;
 };
 
 typedef struct {
